app/requirements.txt
anyio==3.6.2
asgiref==3.3.4
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.5.7
certifi==2021.5.30
chardet==5.0.0
charset-normalizer==2.1.0
click==8.0.1
colorama==0.4.4
configparser==5.2.0
decorator==4.4.2
dnspython==2.1.0
executing==0.8.3
fastapi==0.65.2
future==0.18.2
h11==0.12.0
httptools==0.2.0
huepy==1.2.1
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
iniconfig==1.1.1
instabot==0.117.0
jupyter-core==4.10.0
matplotlib-inline==0.1.3
mock==4.0.3
mongoengine==0.23.1
moviepy==1.0.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
proglog==0.1.10
prompt-toolkit==3.0.29
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.8.2
Pygments==2.12.0
pymongo==3.11.4
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.21.0
rsa==4.8
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.2.0
starlette==0.14.2
Telethon==1.24.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing-extensions==3.10.0.0
urllib3==1.26.9
uvicorn==0.14.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.4
